Calculator is now an object and has input logic
Calc equals ranamed to evaluate
Removed button in regExps
Removed console logs

In task 3 age is checked just with simple if statement by the help of
reusible inRange function.

Used factory function for the calculator

The calculator now has logic for the sybols and negative numbers,
but it doesn't have rules for fractions
to reject more then one dot in a number

Made the required changes:

Calc equals ranamed to evaluate
Removed button in regExps
Removed console logs

In task 3 age is checked just with simple if statement by the help of
reusible inRange function.

Used factory function for the calculator